adding args(input_dir, output_dir, --camera_maker)

+++ Photo-VideoOrganizor.py
@@ -3,15 +3,27 @@ import shutil
 from datetime import datetime
 import sys
 import exifread
+import argparse
 
 
-def main(input_dir, output_dir):
+def main():
+    parser = argparse.ArgumentParser(description='Move specific camara maker photos from source to destination folder.')
+    parser.add_argument('input_dir', type=str, help='Path to the source folder')
+    parser.add_argument('output_dir', type=str, help='Path to the destination folder')
+    parser.add_argument('--camera_maker', type=str, nargs='*', help='Choose a camera maker such samsung or apple')
+    args = parser.parse_args()
+
+    input_dir = args.input_dir
+    output_dir = args.output_dir
+    camera_makers = [maker.lower() for maker in args.camera_maker] if args.camera_maker else []
 
     def get_date_taken(file_path):
+        global camera_maker
         with open(file_path, 'rb') as f:
             tags = exifread.process_file(f, stop_tag='EXIF DateTimeOriginal')
+            camera_maker = str(tags.get('Image Make', '')).lower() if tags.get('Image Make') is not None else None
             date_taken = tags['EXIF DateTimeOriginal']
-            return datetime.strptime(str(date_taken), '%Y:%m')
+            return datetime.strptime(str(date_taken), '%Y:%m'), camera_maker
 
 
     def organize(input_dir, extension, output_dir):
@@ -30,7 +42,10 @@ def main(input_dir, output_dir):
                 os.makedirs(output_path, exist_ok=True)
                 
                 destination_path = os.path.join(output_path, filename)
-                if not os.path.exists(destination_path):
+
+                if camera_makers == []:
+                    shutil.move(file_path, destination_path)
+                elif not os.path.exists(destination_path) and camera_maker in camera_makers:
                     shutil.move(file_path, destination_path)
                 else:
                     print(f"A file with the name '{filename}' already exists.")
@@ -51,4 +66,4 @@ def main(input_dir, output_dir):
 
 
 if __name__ == '__main__':
-    main(sys.argv[1], sys.argv[2])
+    main()
